PLUGIN-1665 fix oauthInfo property
Fix usage of the oauthInfo plugin property. CDAP does not allow
objects within objects in a PluginConfig. This was preventing
the plugins from getting instantiated whenever the oauthInfo
macro was being used.

Refactored classes to pull OAuthInfo into top level fields in each
plugin config. Introduced a SalesforceConnectorInfo class that
contains logic that used to be in the SalesforceConnectorConfig
class, and changed the various plugin configs to return this
new ConnectorInfo class instead of the ConnectorConfig class.